demo/requirements.txt
pandas
opencv-python
Pillow

Keras
tensorflow
scikit-image
image-classifiers
segmentation-models

dash==0.35.1 
dash-html-components==0.13.4
dash-core-components==0.42.1
dash-table==3.1.11
flask